moved parse_rows to sqltypes to avoid import cycle
