+++ go/sqltypes/parse_rows.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package utils
+package sqltypes
 
 import (
 	"fmt"
@@ -24,11 +24,12 @@ import (
 	"strings"
 	"text/scanner"
 
-	"vitess.io/vitess/go/sqltypes"
 	querypb "vitess.io/vitess/go/vt/proto/query"
 )
 
-func ParseRows(input string) ([]sqltypes.Row, error) {
+// ParseRows parses the output generated by fmt.Sprintf("#v", rows), and reifies the original []sqltypes.Row
+// NOTE: This is not meant for production use!
+func ParseRows(input string) ([]Row, error) {
 	type state int
 	const (
 		stInvalid state = iota
@@ -42,8 +43,8 @@ func ParseRows(input string) ([]sqltypes.Row, error) {
 
 	var (
 		scan   scanner.Scanner
-		result []sqltypes.Row
-		row    sqltypes.Row
+		result []Row
+		row    Row
 		vtype  int32
 		st     = stInit
 	)
@@ -75,7 +76,7 @@ func ParseRows(input string) ([]sqltypes.Row, error) {
 				ident := scan.TokenText()
 
 				if ident == "NULL" {
-					row = append(row, sqltypes.NULL)
+					row = append(row, NULL)
 					continue
 				}
 
@@ -101,7 +102,7 @@ func ParseRows(input string) ([]sqltypes.Row, error) {
 				}
 				fallthrough
 			case scanner.Int, scanner.Float:
-				row = append(row, sqltypes.MakeTrusted(sqltypes.Type(vtype), []byte(literal)))
+				row = append(row, MakeTrusted(Type(vtype), []byte(literal)))
 				next = stInValue2
 			}
 		case stInValue2:
@@ -119,14 +120,14 @@ func ParseRows(input string) ([]sqltypes.Row, error) {
 
 type RowMismatchError struct {
 	err       error
-	want, got []sqltypes.Row
+	want, got []Row
 }
 
 func (e *RowMismatchError) Error() string {
 	return fmt.Sprintf("results differ: %v\n\twant: %v\n\tgot:  %v", e.err, e.want, e.got)
 }
 
-func RowsEquals(want, got []sqltypes.Row) error {
+func RowsEquals(want, got []Row) error {
 	if len(want) != len(got) {
 		return &RowMismatchError{
 			err:  fmt.Errorf("expected %d rows in result, got %d", len(want), len(got)),
@@ -163,7 +164,7 @@ func RowsEquals(want, got []sqltypes.Row) error {
 	return nil
 }
 
-func RowsEqualsStr(wantStr string, got []sqltypes.Row) error {
+func RowsEqualsStr(wantStr string, got []Row) error {
 	want, err := ParseRows(wantStr)
 	if err != nil {
 		return fmt.Errorf("malformed row assertion: %w", err)
